refactor: placeSearchService place->ongoing stay 디렉토리로 이동 & SearchService와 FallbackService 분리
